requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-docx
faiss-cpu
openai
